sesion working (Testing for api endpoints pending)

+++ server/modules/ConfigureServer.js
@@ -7,11 +7,25 @@ var fs = require("fs")
 class ConfigureServer {
 
     static async configureSecretFiles(app){
-        // this operation is safe to call various times
-        let serialized_file = await AwsAdmin.read_file_from_bucket("secrets.json", "personal-secret-files");
+        
         let secrets_file_path = path.normalize(__dirname + "/../../credentials/secrets.json")
-        fs.writeFileSync(secrets_file_path, serialized_file);
-        app.locals.secrets = JSON.parse(serialized_file) // share the secrets through the context of the app
+        let serialized_file
+
+        // only download and write to disk if the file does not exists
+        if (!fs.existsSync(secrets_file_path)) {
+
+            serialized_file = await AwsAdmin.read_file_from_bucket("secrets.json", "personal-secret-files");
+            fs.writeFileSync(secrets_file_path, serialized_file);
+
+        // read from disk otherwise
+        } else {
+            serialized_file = fs.readFileSync(secrets_file_path)
+        }
+
+        // share the secrets through the context of the app
+        app.locals = {
+            secrets:JSON.parse(serialized_file)
+        }
     }
 }
